+++ packages/varlet-cli/site/pc/components/AppMobile.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="varlet-site-mobile var-elevation--3">
     <div class="varlet-site-mobile-content">
-      <iframe :src="`./mobile.html#/${componentName}?language=${language}&platform=pc&path=${path}`"></iframe>
+      <iframe :src="`./mobile.html#/${componentName}?language=${language}&platform=pc&replace=${replace}`"></iframe>
     </div>
   </div>
 </template>
@@ -16,7 +16,7 @@ export default {
     language: {
       type: String
     },
-    path: {
+    replace: {
       type: String
     }
   }