+++ packages/varlet-cli/src/config/varlet.config.ts
@@ -1,23 +1,22 @@
-import { pathExistsSync, writeFileSync } from 'fs-extra'
-import hash from 'hash-sum'
-import { resolve } from 'path'
+import { pathExistsSync } from 'fs-extra'
 import { merge } from 'lodash'
-import { VARLET_CONFIG, SITE } from '../shared/constant'
+import { VARLET_CONFIG, SITE_CONFIG } from '../shared/constant'
+import { outputFileSyncOnChange } from '../shared/fsUtils'
 
-function getVarletConfig() {
-  const config = (pathExistsSync(VARLET_CONFIG) && require(VARLET_CONFIG)) || {}
-  const mergedConfig = merge(require('../../varlet.default.config.js'), config)
-  const source = JSON.stringify(mergedConfig, null, 2)
+export function getVarletConfig() {
+  let config: any = {}
+  let defaultConfig: any = {}
 
-  const configId = hash(source)
-  const path = resolve(SITE, `./${configId}.site.config.json`)
-
-  writeFileSync(path, source)
-
-  return {
-    mergedConfig,
-    configId,
+  if (pathExistsSync(VARLET_CONFIG)) {
+    delete require.cache[require.resolve(VARLET_CONFIG)]
+    config = require(VARLET_CONFIG)
   }
-}
+  delete require.cache[require.resolve('../../varlet.default.config.js')]
+  defaultConfig = require('../../varlet.default.config.js')
 
-export const { mergedConfig: varletConfig, configId } = getVarletConfig()
+  const mergedConfig = merge(defaultConfig, config)
+  const source = JSON.stringify(mergedConfig, null, 2)
+  outputFileSyncOnChange(SITE_CONFIG, source)
+
+  return mergedConfig
+}