+++ packages/varlet-cli/src/config/webpack.umd.config.ts
@@ -2,10 +2,11 @@ import merge from 'webpack-merge'
 import { resolve } from 'path'
 import { BASE_CONFIG } from './webpack.base.config'
 import { ES_DIR, UMD_DIR } from '../shared/constant'
-import { varletConfig } from './varlet.config'
+import { getVarletConfig } from './varlet.config'
 import { get } from 'lodash'
 
 export function getUmdConfig() {
+  const varletConfig = getVarletConfig()
   const name: string = get(varletConfig, 'name')
 
   return merge(BASE_CONFIG, {