+++ packages/varlet-cli/src/shared/fsUtils.ts
@@ -1,6 +1,6 @@
 import { parse, extname, resolve } from 'path'
-import { lstatSync, pathExistsSync, readdir, readdirSync } from 'fs-extra'
-import { EXAMPLE_DIR_NAME, SRC_DIR, TESTS_DIR_NAME } from './constant'
+import { ensureFileSync, lstatSync, outputFileSync, pathExistsSync, readdir, readdirSync, readFileSync } from 'fs-extra'
+import { EXAMPLE_DIR_NAME, SCRIPTS_EXTENSIONS, SRC_DIR, TESTS_DIR_NAME } from './constant'
 
 export async function getComponentNames(): Promise<string[]> {
   const srcDir: string[] = await readdir(SRC_DIR)
@@ -33,8 +33,16 @@ export function hasSFC(path: string): boolean {
   return dir.some((filename) => isSFC(resolve(path, filename)))
 }
 
+export function outputFileSyncOnChange(path: string, code: string) {
+  ensureFileSync(path)
+  const content = readFileSync(path, 'utf-8')
+  if (content !== code) {
+    outputFileSync(path, code)
+  }
+}
+
 export function isScript(path: string): boolean {
-  return (pathExistsSync(path) && extname(path) === '.js') || extname(path) === '.ts'
+  return pathExistsSync(path) && SCRIPTS_EXTENSIONS.includes(extname(path))
 }
 
 export function isLess(path: string): boolean {