+++ packages/varlet-ui/babel.config.js
@@ -1,4 +1,3 @@
 module.exports = {
-  presets: ['@babel/preset-env', '@babel/preset-typescript'],
-  plugins: ['@babel/plugin-transform-runtime', '@babel/plugin-transform-typescript'],
+  presets: ['@varlet/cli/preset'],
 }