+++ packages/varlet-ui/src/home/example/index.vue
@@ -37,7 +37,7 @@ export default {
     const logo = ref(config?.logo ?? '')
     const description = ref(config?.mobile?.title ?? {})
     const menu = ref(config?.pc?.menu ?? [])
-    const configComponents = menu.value.filter((item) => !item.isTitle && !item.nonComponent)
+    const configComponents = menu.value.filter((item) => item.type === 2)
     const components = reactive(configComponents)
     const lang = ref('zh-CN')
     const platform = ref('mobile')
@@ -57,7 +57,6 @@ export default {
         query: {
           language: lang.value,
           platform: platform.value,
-          path: component.doc,
         },
       })
     }
@@ -76,6 +75,7 @@ export default {
 
 <style scoped lang="less">
 @import '../../styles/var';
+
 .logo {
   height: 100px;
   padding-top: 30px;