+++ packages/varlet-ui/varlet.config.js
@@ -1,6 +1,7 @@
 module.exports = {
   host: '0.0.0.0',
   port: 8080,
+  useMobile: true,
   pc: {
     menu: [
       {
@@ -8,7 +9,7 @@ module.exports = {
           'zh-CN': '开发指南',
           'en-US': 'Developer guide',
         },
-        isTitle: true,
+        type: 1,
       },
       {
         text: {
@@ -16,7 +17,7 @@ module.exports = {
           'en-US': 'Basic Introduce',
         },
         doc: 'home',
-        nonComponent: true,
+        type: 3,
       },
       {
         text: {
@@ -24,7 +25,7 @@ module.exports = {
           'en-US': 'Quickstart',
         },
         doc: 'quickstart',
-        nonComponent: true,
+        type: 3,
       },
       {
         text: {
@@ -32,7 +33,7 @@ module.exports = {
           'en-US': 'Import on demand',
         },
         doc: 'importOnDemand',
-        nonComponent: true,
+        type: 3,
       },
       {
         text: {
@@ -40,7 +41,7 @@ module.exports = {
           'en-US': 'Browser Adaptation',
         },
         doc: 'browserAdaptation',
-        nonComponent: true,
+        type: 3,
       },
       {
         text: {
@@ -48,7 +49,7 @@ module.exports = {
           'en-US': 'Locale',
         },
         doc: 'locale',
-        nonComponent: true,
+        type: 3,
       },
       {
         text: {
@@ -56,7 +57,7 @@ module.exports = {
           'en-US': 'IDE Code Support',
         },
         doc: 'ide',
-        nonComponent: true,
+        type: 3,
       },
       {
         text: {
@@ -64,14 +65,14 @@ module.exports = {
           'en-US': 'Developer Guide',
         },
         doc: 'developerGuide',
-        nonComponent: true,
+        type: 3,
       },
       {
         text: {
           'zh-CN': '基础组件',
           'en-US': 'Basic Components',
         },
-        isTitle: true,
+        type: 1,
       },
       {
         text: {
@@ -79,6 +80,7 @@ module.exports = {
           'en-US': 'StyleProvider',
         },
         doc: 'style-provider',
+        type: 2,
       },
       {
         text: {
@@ -86,6 +88,7 @@ module.exports = {
           'en-US': 'Button',
         },
         doc: 'button',
+        type: 2,
       },
       {
         text: {
@@ -93,6 +96,7 @@ module.exports = {
           'en-US': 'Cell',
         },
         doc: 'cell',
+        type: 2,
       },
       {
         text: {
@@ -100,6 +104,7 @@ module.exports = {
           'en-US': 'Icon',
         },
         doc: 'icon',
+        type: 2,
       },
       {
         text: {
@@ -107,6 +112,7 @@ module.exports = {
           'en-US': 'Image',
         },
         doc: 'image',
+        type: 2,
       },
       {
         text: {
@@ -114,6 +120,7 @@ module.exports = {
           'en-US': 'Loading',
         },
         doc: 'loading',
+        type: 2,
       },
       {
         text: {
@@ -121,6 +128,7 @@ module.exports = {
           'en-US': 'Chip',
         },
         doc: 'chip',
+        type: 2,
       },
       {
         text: {
@@ -128,6 +136,7 @@ module.exports = {
           'en-US': 'Badge',
         },
         doc: 'badge',
+        type: 2,
       },
       {
         text: {
@@ -135,13 +144,14 @@ module.exports = {
           'en-US': 'Elevation',
         },
         doc: 'styles',
+        type: 2,
       },
       {
         text: {
           'zh-CN': '展示组件',
           'en-US': 'Display Components',
         },
-        isTitle: true,
+        type: 1,
       },
       {
         text: {
@@ -149,6 +159,7 @@ module.exports = {
           'en-US': 'Skeleton',
         },
         doc: 'skeleton',
+        type: 2,
       },
       {
         text: {
@@ -156,6 +167,7 @@ module.exports = {
           'en-US': 'Collapse',
         },
         doc: 'collapse',
+        type: 2,
       },
       {
         text: {
@@ -163,6 +175,7 @@ module.exports = {
           'en-US': 'Layout',
         },
         doc: 'row',
+        type: 2,
       },
       {
         text: {
@@ -170,6 +183,7 @@ module.exports = {
           'en-US': 'Sticky',
         },
         doc: 'sticky',
+        type: 2,
       },
       {
         text: {
@@ -177,6 +191,7 @@ module.exports = {
           'en-US': 'Progress',
         },
         doc: 'progress',
+        type: 2,
       },
       {
         text: {
@@ -184,6 +199,7 @@ module.exports = {
           'en-US': 'List',
         },
         doc: 'list',
+        type: 2,
       },
       {
         text: {
@@ -191,6 +207,7 @@ module.exports = {
           'en-US': 'Swipe',
         },
         doc: 'swipe',
+        type: 2,
       },
       {
         text: {
@@ -198,6 +215,7 @@ module.exports = {
           'en-US': 'Steps',
         },
         doc: 'steps',
+        type: 2,
       },
       {
         text: {
@@ -205,6 +223,7 @@ module.exports = {
           'en-US': 'ImagePreview',
         },
         doc: 'image-preview',
+        type: 2,
       },
       {
         text: {
@@ -212,6 +231,7 @@ module.exports = {
           'en-US': 'Card',
         },
         doc: 'card',
+        type: 2,
       },
       {
         text: {
@@ -219,13 +239,14 @@ module.exports = {
           'en-US': 'Divider',
         },
         doc: 'divider',
+        type: 2,
       },
       {
         text: {
           'zh-CN': '导航组件',
           'en-US': 'Navigation Components',
         },
-        isTitle: true,
+        type: 1,
       },
       {
         text: {
@@ -233,6 +254,7 @@ module.exports = {
           'en-US': 'Tabs',
         },
         doc: 'tabs',
+        type: 2,
       },
       {
         text: {
@@ -240,6 +262,7 @@ module.exports = {
           'en-US': 'IndexBar',
         },
         doc: 'index-bar',
+        type: 2,
       },
       {
         text: {
@@ -247,13 +270,14 @@ module.exports = {
           'en-US': 'AppBar',
         },
         doc: 'app-bar',
+        type: 2,
       },
       {
         text: {
           'zh-CN': '功能指令',
           'en-US': 'Functional Directives',
         },
-        isTitle: true,
+        type: 1,
       },
       {
         text: {
@@ -261,6 +285,7 @@ module.exports = {
           'en-US': 'Ripple',
         },
         doc: 'ripple',
+        type: 2,
       },
       {
         text: {
@@ -268,13 +293,14 @@ module.exports = {
           'en-US': 'Lazy',
         },
         doc: 'lazy',
+        type: 2,
       },
       {
         text: {
           'zh-CN': '反馈组件',
           'en-US': 'Action Components',
         },
-        isTitle: true,
+        type: 1,
       },
       {
         text: {
@@ -282,6 +308,7 @@ module.exports = {
           'en-US': 'Snackbar',
         },
         doc: 'snackbar',
+        type: 2,
       },
       {
         text: {
@@ -289,6 +316,7 @@ module.exports = {
           'en-US': 'ActionSheet',
         },
         doc: 'action-sheet',
+        type: 2,
       },
       {
         text: {
@@ -296,6 +324,7 @@ module.exports = {
           'en-US': 'Dialog',
         },
         doc: 'dialog',
+        type: 2,
       },
       {
         text: {
@@ -303,6 +332,7 @@ module.exports = {
           'en-US': 'PullRefresh',
         },
         doc: 'pull-refresh',
+        type: 2,
       },
       {
         text: {
@@ -310,6 +340,7 @@ module.exports = {
           'en-US': 'Popup',
         },
         doc: 'popup',
+        type: 2,
       },
       {
         text: {
@@ -317,6 +348,7 @@ module.exports = {
           'en-US': 'Menu',
         },
         doc: 'menu',
+        type: 2,
       },
       {
         text: {
@@ -324,6 +356,7 @@ module.exports = {
           'en-US': 'BackTop',
         },
         doc: 'back-top',
+        type: 2,
       },
       {
         text: {
@@ -331,6 +364,7 @@ module.exports = {
           'en-US': 'Countdown',
         },
         doc: 'countdown',
+        type: 2,
       },
       {
         text: {
@@ -338,6 +372,7 @@ module.exports = {
           'en-US': 'Picker',
         },
         doc: 'picker',
+        type: 2,
       },
       {
         text: {
@@ -345,6 +380,7 @@ module.exports = {
           'en-US': 'DatePicker',
         },
         doc: 'date-picker',
+        type: 2,
       },
       {
         text: {
@@ -352,13 +388,14 @@ module.exports = {
           'en-US': 'TimePicker',
         },
         doc: 'time-picker',
+        type: 2,
       },
       {
         text: {
           'zh-CN': '表单组件',
           'en-US': 'Form Components',
         },
-        isTitle: true,
+        type: 1,
       },
       {
         text: {
@@ -366,6 +403,7 @@ module.exports = {
           'en-US': 'Form',
         },
         doc: 'form',
+        type: 2,
       },
       {
         text: {
@@ -373,6 +411,7 @@ module.exports = {
           'en-US': 'Input',
         },
         doc: 'input',
+        type: 2,
       },
       {
         text: {
@@ -380,6 +419,7 @@ module.exports = {
           'en-US': 'Select',
         },
         doc: 'select',
+        type: 2,
       },
       {
         text: {
@@ -387,6 +427,7 @@ module.exports = {
           'en-US': 'Radio',
         },
         doc: 'radio-group',
+        type: 2,
       },
       {
         text: {
@@ -394,6 +435,7 @@ module.exports = {
           'en-US': 'Checkbox',
         },
         doc: 'checkbox-group',
+        type: 2,
       },
       {
         text: {
@@ -401,6 +443,7 @@ module.exports = {
           'en-US': 'Counter',
         },
         doc: 'counter',
+        type: 2,
       },
       {
         text: {
@@ -408,6 +451,7 @@ module.exports = {
           'en-US': 'Switch',
         },
         doc: 'switch',
+        type: 2,
       },
       {
         text: {
@@ -415,6 +459,7 @@ module.exports = {
           'en-US': 'slider',
         },
         doc: 'slider',
+        type: 2,
       },
       {
         text: {
@@ -422,6 +467,7 @@ module.exports = {
           'en-US': 'Rate',
         },
         doc: 'rate',
+        type: 2,
       },
       {
         text: {
@@ -429,8 +475,8 @@ module.exports = {
           'en-US': 'Uploader',
         },
         doc: 'uploader',
+        type: 2,
       },
     ],
   },
-  mobile: {},
 }